Fix libbw compilation on MinGW
At least MinGW supplied with Qt creator defined gmtime_r() and
localtime_r() as macro in pthread.h. That is not cached with the CMake
check. Check it again in the implementation file.

Well, it would be better to change the CMake check, but I don't want to
test again every platform now... :-)

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>